# swagger_ui/openapi_ms1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment Data Service API",
    "description": "Documentação completa das rotas do microserviço investment_data_service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor local"
    }
  ],
  "paths": {
...
    "/admin/investimentos/{id_cliente}": {
      "get": {
        "summary": "Listar investimentos de um cliente",
        "tags": ["Investimentos"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de investimentos"
          }
        }
      }
    },
    "/admin/investidor/{id_cliente}": {
      "get": {
        "summary": "Buscar investidor pelo ID do cliente",
        "tags": ["Investidor"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investidor encontrado"
          },
          "404": {
            "description": "Investidor não encontrado"
          }
        }
      }
    },
...
    "/admin/investimentos/{id_investimento}": {
      "patch": {
        "summary": "Atualizar investimento",
        "tags": ["Investimentos"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": { "type": "string" },
                  "tipo_investimento": { "type": "string" },
                  "valor_investido": { "type": "number" },
                  "data_aplicacao": { "type": "string" },
                  "rentabilidade": { "type": "number" },
                  "ativo": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Investimento atualizado" },
          "404": { "description": "Investimento não encontrado" }
        }
      },

      "delete": {
        "summary": "Deletar investimento",
        "tags": ["Investimentos"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Investimento deletado" },
          "404": { "description": "Investimento não encontrado" }
        }
      }
    },
...
  },  
...
}